Add toolchain configuration for RISC-V Nios V/m
This commit introduces a new CMake file for the RISC-V toolchain, setting up the necessary configurations for cross-compilation. Key settings include the specification of compilers, flags for debug and release builds, and additional compile and link options tailored for the RISC-V architecture. The configuration also includes definitions for logging and JTAG UART settings.

Signed-off-by: Schubert Anselme <schubert@anselm.es>